Fixed more Warnings, though VS2022 goes from saying I have 0, to several for seemingly no reason.
Window Icon now chooses the correct one based on DPI settings.

+++ src/main.cpp
@@ -74,6 +74,6 @@ int CALLBACK WinMain(HINSTANCE hInstance, HINSTANCE, LPSTR, int) {
 
     return static_cast<unsigned int>(retVal);
 #else
-    return retVal;    
+    return static_cast<unsigned int>(retVal);
 #endif // defined(_MSC_VER) && defined(_DEBUG)
 }
